transform adblock-banner-config.spec.js to ES6 module
